Refactored reconstructor out into tree_matcher. Functionality should stay the same.
